Added ProjectDetails

+++ DevCard.MVC/Models/Service.cs
@@ -2,13 +2,13 @@
 {
     public class Service
     {
-        public Service(int id, string name)
+        public Service(long id, string name)
         {
             Id = id;
             Name = name;
         }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
 
     }
